pyproject.toml
[project]
name = "agentsolution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=42.0.0",
    "dspy>=3.1.3",
    "dspy-ai>=3.1.3",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "kafka-python>=2.3.0",
    "langchain-core>=1.2.18",
    "langgraph>=1.1.2",
    "langsmith>=0.7.17",
    "neo4j>=6.1.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pika>=1.3.2",
    "prometheus-client>=0.24.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "strands-agents[otel]>=1.30.0",
    "strands-agents-tools>=0.2.22",
    "weaviate-client>=4.20.4",
]
